+++ src/pages/auth/signin.tsx
@@ -12,8 +12,8 @@ import 'firebase/auth';
 import initFirebase from '../../util/auth/initFirebase';
 import Link from 'next/link';
 import { GetServerSideProps } from 'next';
-import { apiGetSession } from '../api/auth/get_session';
 import { AuthCircular } from '../../app/components/AuthCircular';
+import { apiGetSession } from '../../util/db/apiGetSession';
 
 export const useStyles = (isTabletPortrait: boolean) =>
   makeStyles((theme: Theme) => {
